backend_stub/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
google-generativeai>=0.8.0
openai>=1.0.0
python-dotenv>=1.0.0